feat: #1670 register move to folder
