+++ packages/marketplace/src/components/pages/register/index.ts
@@ -1,0 +1,1 @@
+export { default } from './register'
