+++ packages/marketplace/src/components/ui/admin-approval-modal/approve-revision-modal.tsx
@@ -3,7 +3,7 @@ import { useSelector, useDispatch } from 'react-redux'
 import { ApproveModel } from '@reapit/foundations-ts-definitions'
 import { Button, Modal, ModalProps, ModalBody, SubTitleH6, ModalFooter, Form, Formik } from '@reapit/elements'
 import { approveRevision } from '@/actions/revision-detail'
-import CallToAction from './call-to-action'
+import CallToAction from '../call-to-action'
 import { selectAppRevisionDetail } from '@/selector/app-revisions'
 import { selectLoginIdentity } from '@/selector/auth'
 import { Dispatch } from 'redux'
