+++ packages/marketplace/src/components/ui/admin-approval-modal/decline-revision-modal.tsx
@@ -1,30 +1,14 @@
 import * as React from 'react'
-import * as Yup from 'yup'
 import { Dispatch } from 'redux'
 import { useDispatch, useSelector } from 'react-redux'
 import { RejectRevisionModel } from '@reapit/foundations-ts-definitions'
 import { Button, TextArea, Modal, ModalProps, ModalFooter, ModalBody, Form, Formik } from '@reapit/elements'
 import { declineRevision } from '@/actions/revision-detail'
-import CallToAction from './call-to-action'
+import CallToAction from '../call-to-action'
 import { selectAppRevisionDetail } from '@/selector/app-revisions'
 import { selectLoginIdentity } from '@/selector/auth'
-import { FormFieldInfo } from '@reapit/utils'
-import errorMessages from '@/constants/error-messages'
-
-export type FieldKey = 'rejectionReasonField'
-
-export const formFields: Record<FieldKey, FormFieldInfo> = {
-  rejectionReasonField: {
-    name: 'rejectionReason',
-    label: 'Rejection reason',
-  },
-}
-
-export const validationSchema = Yup.object().shape({
-  [formFields.rejectionReasonField.name]: Yup.string()
-    .trim()
-    .required(errorMessages.FIELD_REQUIRED),
-})
+import { validationSchemaDeclineModal as validationSchema } from './validation-schema'
+import { formFieldsDeclineModal as formFields } from './form-fields'
 
 export type DeclineRevisionModalProps = Pick<ModalProps, 'visible' | 'afterClose'> & {
   onDeclineSuccess: () => void
