+++ packages/marketplace/src/components/ui/admin-approval-modal/form-fields.ts
@@ -1,0 +1,9 @@
+import { FormFieldInfo } from '@reapit/utils'
+export type FieldKeyDeclineModal = 'rejectionReasonField'
+
+export const formFieldsDeclineModal: Record<FieldKeyDeclineModal, FormFieldInfo> = {
+  rejectionReasonField: {
+    name: 'rejectionReason',
+    label: 'Rejection reason',
+  },
+}
