+++ packages/marketplace/src/components/ui/admin-approval-modal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './admin-approval-modal'
